Refactor so functions can be shared between `edit` and `save`
