+++ src/index.js
@@ -5,7 +5,8 @@ import { registerBlockType, getBlockDefaultClassName } from '@wordpress/blocks';
 import { createElement as el } from '@wordpress/element';
 import Masonry from 'react-masonry-component';
 import classnames from 'classnames';
-import edit from './edit.js';
+import createMasonryBrick from './utils';
+import edit from './edit';
 
 registerBlockType( 'sc/double-masonry', {
     title: 'Masonry Gallery',
@@ -26,13 +27,17 @@ registerBlockType( 'sc/double-masonry', {
     save( { attributes } ) {
         const className = getBlockDefaultClassName( 'sc/double-masonry' );
         const hasImages = !! attributes.gallery.length;
+        const masonry = el( Masonry,
+                            {
+                                ref: function( c ) { this.masonry = this.masonry || c.masonry; }.bind( this ),
+                                className: classnames( className, 'my-8' )
+                            },
+                            el( 'div', { className: `${className}__sizer` } ),
+                            attributes.gallery.map( img => {
+                                return createMasonryBrick( img, className, attributes );
+                            } )
+                          );
 
-        const list = el( Masonry, {
-            className: classnames( 'my-8' )
-        }, attributes.gallery.map( img => {
-            return el( 'div', { key: img.id, className: classnames( `${className}__item`, [ 'w-1/3', 'border', 'p-4' ] ) }, ( img.caption || img.url ) )
-        } ) );
-
-        return ( hasImages && list );
+        return ( hasImages && masonry );
     }
 } );
